+++ NPCWandering/Assets/BehaviorTrees.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 74a9793c50f662a489f38189d0c9bd68
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
