+++ NPCWandering/Assets/BehaviorTrees/NPCWander.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61d91722b8ef6e245ab89eddb3179532
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
