+++ NPCWandering/Assets/Scene - Tutorials/00 - UIDialog.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e6beb033284bfec4d9f143a6326c44a0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
