+++ NPCWandering/Assets/Scene - Tutorials/01 - BPWandering.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 76641c59f83fee5419e8f82424a7fc81
-NativeFormatImporter:
+guid: d930c10e9e8092a47b55ab2e1138a125
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 23800000
   userData: 
   assetBundleName: 
   assetBundleVariant:
