+++ NPCWandering/Assets/Scene - Tutorials/02 - BPReceiveDamage.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7a487df25b19f6e45ab1b0e64fbf25a1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
