+++ NPCWandering/Assets/Scene - Tutorials/02 - BPReceiveDamage/BPReceiveDamage.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c465b6df81fb6794cb1fdcf9db0035db
+guid: 2089acac6a8540140911d76e2de0f2e1
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
