+++ NPCWandering/Assets/Scene - Tutorials/02 - BPReceiveDamage/BPReceiveDamage.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b942785658b2c37459841f1b933f8ff6
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
