+++ NPCWandering/Assets/Scene - Tutorials/02 - BPReceiveDamage/BPReceiveDamage/NavMesh-navmesh.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5bc345adfd864342a7c70d5bba764ae
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
