+++ NPCWandering/Assets/Scripts/ReceiveDamage.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 73ba847bc1459744581a2d656dd54b98
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
