+++ NPCWandering/Assets/Scripts/ReceiveDamage/ChaseTarget.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb02f93c02e5ba14aaf9293ad97f8722
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
