+++ NPCWandering/Assets/Scripts/ReceiveDamage/NPCLeashBreaks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 540555bd91b493b41b50d557b0803ac1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
