+++ NPCWandering/Assets/Scripts/ReceiveDamage/NPCReceiveDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5bfd581005a7784418d350964f8d3cc7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
