+++ NPCWandering/Assets/Scripts/UIDialog.meta
