+++ NPCWandering/Assets/Scripts/UIDialog/NPCTrigger.cs.meta
