+++ NPCWandering/Assets/Scripts/UIDialog/VEDialogController.cs.meta
