+++ NPCWandering/Assets/Scripts/UIDialog/VEDialogReader.cs.meta
